+++ src/main/java/com/otus/questionnaire/QuestionaryApplication.java
@@ -1,6 +1,6 @@
 package com.otus.questionnaire;
 
-import com.otus.questionnaire.service.QuestionnaireService;
+import com.otus.questionnaire.quesrionservice.QuestionnaireService;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.PropertySource;
