+++ src/main/java/com/otus/questionnaire/quesrionservice/QuestionnaireService.java
@@ -1,0 +1,5 @@
+package com.otus.questionnaire.quesrionservice;
+
+public interface QuestionnaireService {
+    void questionInit();
+}
