+++ src/main/java/com/otus/questionnaire/quizanswersservice/AnswersReaderImpl.java
@@ -1,0 +1,43 @@
+package com.otus.questionnaire.quizanswersservice;
+
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Scanner;
+
+@Service
+public class AnswersReaderImpl implements AnswerReader {
+
+    private final Scanner scanner;
+
+    public AnswersReaderImpl() {
+        this.scanner = new Scanner(System.in);
+    }
+
+    @Override
+    public String getNames() {
+        String name;
+        while ((name = scanner.nextLine()).isEmpty()) {
+            System.out.println("Please, enter your name");
+        }
+        return name;
+    }
+
+    @Override
+    public List<String> getQuizAnswers(
+            Map<String, List<String>> questionBase,
+            Map<String, String> answerBase) {
+        List<String> answers = new ArrayList<>();
+        questionBase.forEach((key, value) -> {
+            System.out.printf("\n%s?", key);
+            System.out.printf("\nChooze: %s or %s\n", value.get(0), value.get(1));
+            String response = scanner.next();
+            if (response.equalsIgnoreCase(answerBase.get(key))) {
+                answers.add(key + ": " + response);
+            }
+        });
+        return answers;
+    }
+}
